// src/generated/api/model/createUserCommand.ts
/**
 * Hello Doctor Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateUserCommand { 
    email: string | null;
    firstname: string | null;
    lastname: string | null;
    phoneNumber: string | null;
    password: string | null;
    role: string | null;
}